File->Submit to Source Control fails to display Submit Files dialog
In the process of executing the Submit to Source Control command
UnrealEd attempts to obtain source control status info for built-in
assets [1] (a.k.a. engine content). The plugin converts all asset paths to
be relative to the repository directory before passing them on to hg.
If the user creates their project on a different drive to the one the
engine is installed on then the asset paths for built-in content can't
be converted to be relative to the repository directory.

Previously when the plugin failed to convert an asset path to a relative
path it immediately aborted the "update status" operation. This fix
removes the all-or-nothing requirement such the "update status"
operation proceeds so long as at least one valid relative path can be
deduced.

[1] This is apparently by design: https://answers.unrealengine.com/questions/34626/perforce-not-under-clients-root.html